autofield/requirements.txt
pyserial>=3.5
pyubx2>=1.2.44
PyYAML>=6.0.1
